Move retry up a level
